renames track pages and associated tests to experiment pages
